Add runtime cloth/rope configuration
- Modifies the DeformableObjectConfiguration base class to be much more
generic, allowing for initialization of both rope and cloth derived
classes while also allowing for the CDCPD_Moveit_Node to keep a smart
pointer to the base class. This cleans the code up much more and allows
for the runtime choice of tracking cloth or rope.
- Unfortunately, this does rely on rosparams to do this runtime
choice between cloth and rope tracking. This can be addressed in the
future.
- Adds skeleton code for warping cloth template initialization by an
affine transform. This will be helpful when initializing cloth tracking
(potentially multi-template tracking) based off of segmentation of the
incoming point cloud or RGBD data.
- Updates the readme to assist with configuring cloth versus rope
tracking.
- Additionally addresses #20 and #28