Clean up code
- Cleans up some of the includes in CDCPD_Moveit_Node and changes how
the sdf_filename parameter is loaded since it required a different node
handler.
- Working implementation! I've got this to work but need to set some ros
params via the shell since those rosparams rely on containing a full
srdf and urdf file in the rosparam, not just the file names. I need to
figure out how to do this in a simpler way.
- Still does not handle cases where the collision sphere is inside of an
obstacle.