Add last bits of work before abandoning
- Adds the last chunks of work (partial thoughts mostly) before this
push for the IROS 2022 cloth corner perception challenge was abandoned.
- I'm keeping this around as it might come in handy later, but the
problem is that this initiative was approached as handling
multi-template tracking with multiple CDCPD instances. This is NOT how
we want to proceed as this completely disregards template-template
interactions and acts like neither template can see or interact with one
another.
- New approach (which will be on a completely different branch) will
attempt to run all templates at once, using the tracking to aid in
segmentation clustering (where this means that we'll try to use tracking
to separate template instances from one another).